saving messages in chats fix bugs

+++ anon_news/apps/chat/views.py
@@ -1,8 +1,9 @@
-from django.contrib.auth.decorators import user_passes_test
-from django.shortcuts import render, get_object_or_404
+from django.contrib.auth.decorators import user_passes_test, login_required
+from django.shortcuts import render, get_object_or_404, redirect
+from django.urls import reverse
 from django.views.generic import ListView
 
-from apps.chat.models import Chat
+from apps.chat.models import Chat, Message
 
 
 # Create your views here.
@@ -28,12 +29,24 @@ def index(request):
 
 
 def room(request, room_name):
-    return render(request, "chat/room.html", {"room_name": room_name})
+    # Получить объект чата по имени комнаты
+    chat = Chat.objects.get(name=room_name)
+
+    # Получите все сообщения для данного чата
+    messages = Message.objects.filter(chat=chat)
+
+    context = {
+        "room_name": room_name,
+        "messages": messages,
+    }
+    return render(request, "chat/room.html", context)
 
 
-# СЮДА НУЖНО ДОБАВИТЬ БЕЗОПАСНОСТЬ!!!
+@login_required
 def delete_chat(request, chat_name):
     chat = get_object_or_404(Chat, name=chat_name)
+    if request.user not in chat.participants.all():
+        return redirect('forbidden')
     chat.delete()
     queryset = request.user.chats.all()
     context = {
